add new chord tool
